♻️ refactor : InputView.js 에 있던 비즈니스 로직 이동 및 수정
InputVive.js 는 UI 로직을 담당하기에 존재하던 비즈니스 로직을 따로 retryInput.js 로 이동.
또한 Validate의 필요없는 필드 삭제
