+++ src/Validate.js
@@ -1,17 +1,13 @@
 import { error, magicNumber, uiConstants } from './constants/index.js';
 
 export default class Validate {
-  #date;
-
-  #menu;
-
-  constructor(obj, type) {
+  excute(obj, type) {
     switch (type) {
       case uiConstants.DATE:
-        this.#date = this.#dateValidate(obj);
+        this.#dateValidate(obj);
         break;
       case uiConstants.MENU:
-        // this.#menu = this.#menuValidate(obj);
+        // this.#menu = obj;
         break;
     }
   }
@@ -24,20 +20,22 @@ export default class Validate {
 
   #menuValidate(obj) {}
 
+  // --------------
+  // 유효성 검사 메서드모음
   // 숫자인지 아닌지
   #isNumber(obj) {
-    if (isNaN(obj)) throw new Error(error.WRONG_NUMBER);
+    if (isNaN(obj)) throw new Error(error.WRONG_DATE);
   }
 
   // 공백인지 아닌지 체크
   #isEmpty(obj) {
     const regExp = /\s/;
-    if (!obj || regExp.test(obj)) throw new Error(error.WRONG_NUMBER);
+    if (!obj || regExp.test(obj)) throw new Error(error.WRONG_DATE);
   }
 
   // 숫자가 조건에 맞는지 (1~31), 0인지 아닌지도 검사가 된다.
   #isInRange(num) {
     if (num < magicNumber.FIRST_DAY || num > magicNumber.END_DAY)
-      throw new Error(error.WRONG_NUMBER);
+      throw new Error(error.WRONG_DATE);
   }
 }
