+++ src/constants/error.js
@@ -1,5 +1,6 @@
 const error = {
-  WRONG_NUMBER: '[ERROR] 유효하지 않은 날짜입니다. 다시 입력해 주세요.',
+  WRONG_DATE: '[ERROR] 유효하지 않은 날짜입니다. 다시 입력해 주세요.',
+  WRONG_MENU: '[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.',
 };
 
 export default error;
