+++ src/utils/retryInput.js
@@ -1,6 +1,6 @@
 import { Console } from '@woowacourse/mission-utils';
+import { OutputView } from '../view/index.js';
 import { uiConstants } from '../constants/index.js';
-import { OutputView } from './index.js';
 import Validate from '../Validate.js';
 
 const retry = {
@@ -9,7 +9,8 @@ const retry = {
 
     try {
       result = await this.createType(type);
-      new Validate(result, type);
+      const validate = new Validate();
+      validate.excute(result, type);
     } catch (error) {
       OutputView.printError(error.message);
       result = await this.createInput(type);
@@ -31,16 +32,4 @@ const retry = {
   },
 };
 
-const InputView = {
-  async readDate() {
-    const input = await retry.createInput(uiConstants.DATE);
-    return input;
-  },
-
-  async orderMenu() {
-    const input = await retry.createInput(uiConstants.MENU);
-    return input;
-  },
-};
-
-export default InputView;
+export default retry;
